Create-JWT-service: move JwtTokenFilter to security/filter package
